requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
langchain-groq
faiss-cpu
sentence-transformers